+++ @xen-orchestra/lite/src/views/new-vm/NewVmView.vue
@@ -118,7 +118,15 @@
                 <VtsInputWrapper :label="$t('new-vm.name')">
                   <UiInput v-model="vmState.name" accent="brand" />
                 </VtsInputWrapper>
-                <!-- <UiInput v-model="vmState.tags" :label-icon="faTags" accent="brand" :label=" $t('tags')" /> -->
+                <VtsInputWrapper :label="$t('tags')" :icon="faTags">
+                  <!-- TODO Change input text into select when Thierry's component is available -->
+                  <UiInput v-model="vmState.tag" accent="brand" @keydown.enter.prevent="addTag" />
+                </VtsInputWrapper>
+                <div v-if="vmState.tags.length > 0" class="chips">
+                  <UiChip v-for="(tag, index) in vmState.tags" :key="index" accent="info" @remove="removeTag(index)">
+                    {{ tag }}
+                  </UiChip>
+                </div>
                 <VtsInputWrapper :label="$t('boot-firmware')">
                   <FormSelect v-model="vmState.boot_firmware">
                     <option v-for="boot in bootFirmwares" :key="boot" :value="boot">
@@ -150,7 +158,16 @@
                 <UiTextarea v-model="vmState.description" accent="brand">
                   {{ $t('new-vm.description') }}
                 </UiTextarea>
-                <!-- <UiInput v-model="vmState.affinity_host" accent="brand" :label="$t('affinity-host')" /> -->
+                <VtsInputWrapper :label="$t('affinity-host')">
+                  <div class="affinity-host">
+                    <FormSelect v-model="vmState.affinity_host" class="select">
+                      <option :value="OPAQUE_REF.EMPTY">{{ $t('none') }}</option>
+                      <option v-for="host in hosts" :key="host?.$ref" :value="host?.$ref">
+                        {{ host?.name_label }}
+                      </option>
+                    </FormSelect>
+                  </div>
+                </VtsInputWrapper>
               </div>
             </div>
             <!-- MEMORY SECTION -->
@@ -397,6 +414,7 @@ import FormSelect from '@/components/form/FormSelect.vue'
 
 // XenAPI Store imports
 import type { XenApiVdi, XenApiVm } from '@/libs/xen-api/xen-api.types'
+import { useHostStore } from '@/stores/xen-api/host.store.ts'
 import { useNetworkStore } from '@/stores/xen-api/network.store'
 import { usePifStore } from '@/stores/xen-api/pif.store'
 import { usePoolStore } from '@/stores/xen-api/pool.store'
@@ -419,6 +437,7 @@ import UiButtonIcon from '@core/components/ui/button-icon/UiButtonIcon.vue'
 import UiCard from '@core/components/ui/card/UiCard.vue'
 import UiCheckbox from '@core/components/ui/checkbox/UiCheckbox.vue'
 import UiCheckboxGroup from '@core/components/ui/checkbox-group/UiCheckboxGroup.vue'
+import UiChip from '@core/components/ui/chip/UiChip.vue'
 import UiHeadBar from '@core/components/ui/head-bar/UiHeadBar.vue'
 import UiInput from '@core/components/ui/input/UiInput.vue'
 import UiRadioButton from '@core/components/ui/radio-button/UiRadioButton.vue'
@@ -437,6 +456,7 @@ import {
   faMicrochip,
   faNetworkWired,
   faPlus,
+  faTags,
   faTrash,
 } from '@fortawesome/free-solid-svg-icons'
 
@@ -451,6 +471,7 @@ import { useRouter } from 'vue-router'
 const { templates } = useVmStore().subscribe()
 const { pool } = usePoolStore().subscribe()
 const { records: srs, vdiIsosBySrName } = useSrStore().subscribe()
+const { records: hosts } = useHostStore().subscribe()
 const { records: networks, getByOpaqueRef: getNetworkByOpaqueRef } = useNetworkStore().subscribe()
 const { getByOpaqueRef: getVbdByOpaqueRef } = useVbdStore().subscribe()
 const { getByOpaqueRef: getVdiByOpaqueRef } = useVdiStore().subscribe()
@@ -472,8 +493,9 @@ const vmState = reactive<VmState>({
   description: '',
   toggle: false,
   installMode: '',
+  tag: '',
   tags: [],
-  affinity_host: '',
+  affinity_host: OPAQUE_REF.EMPTY,
   boot_firmware: '',
   new_vm_template: undefined,
   boot_vm: true,
@@ -536,6 +558,20 @@ const addStorageEntry = () => {
 // const removeSshKey = (index: number) => {
 //   vmState.sshKeys.splice(index, 1)
 // }
+
+const addTag = () => {
+  const tag = vmState.tag.trim()
+
+  if (tag && !vmState.tags.includes(tag)) {
+    vmState.tags.push(tag)
+  }
+
+  vmState.tag = ''
+}
+
+const removeTag = (index: number) => {
+  vmState.tags.splice(index, 1)
+}
 
 const deleteItem = <T,>(array: T[], index: number) => {
   array.splice(index, 1)
@@ -698,27 +734,28 @@ const getVdis = (template: XenApiVm) => {
 }
 
 const getExistingVdis = (template: XenApiVm) => {
-  const existingVdisArray = [] as Vdi[]
-
-  template.VBDs.forEach(vbdRef => {
+  return template.VBDs.reduce<Vdi[]>((acc, vbdRef) => {
     const vbd = getVbdByOpaqueRef(vbdRef)
 
     if (!vbd || vbd.type === 'CD') {
-      return
+      return acc
     }
 
     const vdi = getVdiByOpaqueRef(vbd.VDI)
-    if (!vdi) return
 
-    existingVdisArray.push({
+    if (!vdi) {
+      return acc
+    }
+
+    acc.push({
       name_label: vdi.name_label,
       name_description: vdi.name_description,
       size: bytesToGiB(vdi.virtual_size),
-      SR: vdi.SR ?? defaultSr.value,
+      SR: vdi.SR,
     })
-  })
 
-  return existingVdisArray
+    return acc
+  }, [])
 }
 
 const onTemplateChange = () => {
@@ -733,6 +770,8 @@ const onTemplateChange = () => {
     memory_dynamic_max,
     other_config,
     platform,
+    tags,
+    affinity,
   } = template
 
   Object.assign(vmState, {
@@ -742,7 +781,9 @@ const onTemplateChange = () => {
     vCPU: VCPUs_at_startup,
     ram: memory_dynamic_max,
     vdis: getVdis(template),
+    tags,
     topology: platform['cores-per-socket'] ?? null,
+    affinity_host: affinity,
     existingVdis: getExistingVdis(template),
     networkInterfaces: getExistingInterface(template),
   })
@@ -768,7 +809,7 @@ const vmCreationParams = computed(() => ({
   autoPoweron: vmState.auto_power,
   bootAfterCreate: vmState.boot_vm,
   copyHostBiosStrings: vmState.boot_firmware !== 'uefi' && !templateHasBiosStrings.value && vmState.copyHostBiosStrings,
-  hvmBootFirmware: vmState.boot_firmware,
+  hvmBootFirmware: vmState.boot_firmware ?? 'bios',
   coresPerSocket: vmState.topology,
   tags: vmState.tags,
   cloudConfig: '',
@@ -821,6 +862,8 @@ const _createVm = async ($defer: Defer) => {
     await Promise.all([
       xapi.vm.setNameLabel(vmRefs, vmCreationParams.value.name_label),
       xapi.vm.setNameDescription(vmRefs, vmCreationParams.value.name_description),
+      xapi.vm.setTags(vmRefs, vmCreationParams.value.tags),
+      xapi.vm.setAffinityHost(vmRefs, vmCreationParams.value.affinityHost),
       xapi.vm.setMemory(vmRefs, vmCreationParams.value.memory),
       xapi.vm.setAutoPowerOn(vmRefs[0], vmCreationParams.value.autoPoweron),
       xapi.vm.setCoresPerSocket(vmRefs[0], vmCreationParams.value.coresPerSocket),
@@ -998,6 +1041,23 @@ const createVM = defer(_createVm)
         flex-direction: column;
         gap: 2.5rem;
         width: 40%;
+
+        .affinity-host {
+          display: flex;
+          align-items: center;
+          gap: 0.8rem;
+
+          .container {
+            width: 100%;
+          }
+        }
+      }
+
+      .chips {
+        display: flex;
+        flex-wrap: wrap;
+        gap: 0.5rem;
+        margin-block-end: 1rem;
       }
     }