+++ @xen-orchestra/web-core/lib/components/tree/VtsTreeItem.vue
@@ -1,11 +1,13 @@
 <template>
-  <li class="vts-tree-item">
+  <li class="vts-tree-item" @click="handleClick()">
     <slot />
     <slot v-if="expanded" name="sublist" />
   </li>
 </template>
 
 <script lang="ts" setup>
+import { useSidebarStore } from '@core/stores/sidebar.store'
+import { useUiStore } from '@core/stores/ui.store'
 import { IK_TREE_ITEM_EXPANDED, IK_TREE_ITEM_HAS_CHILDREN } from '@core/utils/injection-keys.util'
 import { onBeforeMount, onBeforeUpdate, provide, ref, toRef, useSlots } from 'vue'
 
@@ -18,11 +20,19 @@ defineSlots<{
   sublist?(): any
 }>()
 
+const sidebar = useSidebarStore()
+const uiStore = useUiStore()
 const hasChildren = ref(false)
 
 const updateHasChildren = () => {
   const { sublist } = useSlots()
   hasChildren.value = sublist !== undefined
+}
+
+const handleClick = () => {
+  if (uiStore.isMobile) {
+    sidebar.toggleExpand(false)
+  }
 }
 
 onBeforeMount(() => updateHasChildren())